@@ -3,19 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.18.0",
-    "classnames": "^2.2.6",
-    "node-sass": "^4.10.0",
+    "axios": "^0.26.0",
+    "classnames": "^2.3.1",
+    "node-sass": "^6.0.1",
     "normalize": "^0.3.1",
-    "normalize.css": "^8.0.1",
-    "react": "^16.6.3",
-    "react-app-polyfill": "^0.1.3",
-    "react-dom": "^16.6.3",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.1",
-    "react-typist": "^2.0.4",
-    "validator": "^10.9.0"
+    "normalize.css": "8.0.1",
+    "react": "^17.0.2",
+    "react-app-polyfill": "^3.0.0",
+    "react-dom": "^17.0.2",
+    "react-router": "^6.2.1",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "^5.0.0",
+    "react-typist": "^2.0.5",
+    "validator": "^13.7.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -32,5 +32,8 @@
     "not ie <= 11",
     "not op_mini all"
   ],
-  "proxy": "http://localhost:5000"
+  "proxy": "http://localhost:5000",
+  "engines": {
+    "node": ">=16.0.0"
+  }
 }
