@@ -1,5 +1,5 @@
 import React, { Fragment } from "react";
-import { Router, Route, Switch } from "react-router-dom";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import { createBrowserHistory } from "history";
 import DashboardPage from "../components/DashboardPage";
 import NotFoundPage from "../components/NotFoundPage";
@@ -9,10 +9,10 @@ export const history = createBrowserHistory();
 const AppRouter = () => (
   <Router history={history}>
     <Fragment>
-      <Switch>
-        <Route path="/" component={DashboardPage} exact={true} />
-        <Route component={NotFoundPage} />
-      </Switch>
+      <Routes>
+        <Route path="/" element={<DashboardPage/>} />
+        <Route path="*" element={<NotFoundPage/>} />
+      </Routes>
     </Fragment>
   </Router>
 );
